Move `option` package under `generate`
