+++ java/src/main/kotlin/io/spine/validation/java/JavaValidationRenderer.kt
@@ -39,10 +39,10 @@ import io.spine.protodata.render.SourceFileSet
 import io.spine.tools.code.Java
 import io.spine.validation.java.generate.MessageValidationCode
 import io.spine.validation.java.generate.ValidationCodeInjector
-import io.spine.validation.java.option.DistinctGenerator
-import io.spine.validation.java.option.GoesGenerator
-import io.spine.validation.java.option.PatternGenerator
-import io.spine.validation.java.option.RequiredGenerator
+import io.spine.validation.java.generate.option.DistinctGenerator
+import io.spine.validation.java.generate.option.GoesGenerator
+import io.spine.validation.java.generate.option.PatternGenerator
+import io.spine.validation.java.generate.option.RequiredGenerator
 import io.spine.validation.java.rule.RuleGenerator
 
 /**
