+++ java/src/main/kotlin/io/spine/validation/java/generate/option/PatternFieldGenerator.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java.option
+package io.spine.validation.java.generate.option
 
 import io.spine.base.FieldPath
 import io.spine.option.PatternOption
